chore: move Source Card to its own file
