+++ app/common/renderer/components/Inspector/Inspector.jsx
@@ -1,14 +1,11 @@
 import {
   CheckCircleOutlined,
   CloseCircleOutlined,
-  CodeOutlined,
-  CopyOutlined,
   DownloadOutlined,
-  FileTextOutlined,
   PlusSquareOutlined,
   SelectOutlined,
 } from '@ant-design/icons';
-import {Button, Card, Modal, Space, Spin, Splitter, Switch, Tabs, Tooltip} from 'antd';
+import {Button, Modal, Space, Spin, Splitter, Switch, Tabs, Tooltip} from 'antd';
 import _ from 'lodash';
 import {useEffect, useRef, useState} from 'react';
 import {useNavigate} from 'react-router';
@@ -21,7 +18,6 @@ import {
   MJPEG_STREAM_CHECK_INTERVAL,
   SESSION_EXPIRY_PROMPT_TIMEOUT,
 } from '../../constants/session-inspector';
-import {copyToClipboard} from '../../polyfills';
 import {downloadFile} from '../../utils/file-handling';
 import Commands from './Commands.jsx';
 import GestureEditor from './GestureEditor.jsx';
@@ -35,12 +31,6 @@ import SessionInfo from './SessionInfo.jsx';
 import Source from './Source.jsx';
 
 const {SELECT, TAP_SWIPE} = SCREENSHOT_INTERACTION_MODE;
-
-const downloadXML = (sourceXML) => {
-  const href = 'data:application/xml;charset=utf-8,' + encodeURIComponent(sourceXML);
-  const filename = `app-source-${new Date().toJSON()}.xml`;
-  downloadFile(href, filename);
-};
 
 const downloadScreenshot = (screenshot) => {
   const href = `data:image/png;base64,${screenshot}`;
@@ -61,7 +51,6 @@ const Inspector = (props) => {
     setUserWaitTimeout,
     showKeepAlivePrompt,
     keepSessionAlive,
-    sourceXML,
     visibleCommandResult,
     serverDetails,
     isUsingMjpegMode,
@@ -69,7 +58,6 @@ const Inspector = (props) => {
     toggleShowCentroids,
     showCentroids,
     isGestureEditorVisible,
-    toggleShowAttributes,
     isSourceRefreshOn,
     windowSize,
     t,
@@ -287,43 +275,7 @@ const Inspector = (props) => {
               children: (
                 <Splitter>
                   <Splitter.Panel collapsible defaultSize="50%" min="25%" max="80%">
-                    <Card
-                      title={
-                        <span>
-                          <FileTextOutlined /> {t('App Source')}{' '}
-                        </span>
-                      }
-                      extra={
-                        <span>
-                          <Tooltip title={t('Toggle Attributes')}>
-                            <Button
-                              type="text"
-                              id="btnToggleAttrs"
-                              icon={<CodeOutlined />}
-                              onClick={toggleShowAttributes}
-                            />
-                          </Tooltip>
-                          <Tooltip title={t('Copy XML Source to Clipboard')}>
-                            <Button
-                              type="text"
-                              id="btnSourceXML"
-                              icon={<CopyOutlined />}
-                              onClick={() => copyToClipboard(sourceXML)}
-                            />
-                          </Tooltip>
-                          <Tooltip title={t('Download Source as .XML File')}>
-                            <Button
-                              type="text"
-                              id="btnDownloadSourceXML"
-                              icon={<DownloadOutlined />}
-                              onClick={() => downloadXML(sourceXML)}
-                            />
-                          </Tooltip>
-                        </span>
-                      }
-                    >
-                      <Source {...props} />
-                    </Card>
+                    <Source {...props} />
                   </Splitter.Panel>
                   <Splitter.Panel collapsible>
                     <SelectedElementContainer {...props} />
